Use the project's full display name in the message
Previously it was using display name. The difference is for projects
nested in folders. getDisplayName returns the last segment there,
whereas getFullDisplayName includes all the parent folders as well,
separated by ». For projects not in folders, they're identical.

We have several builds with the same final segment name
that are only different based on the folder they're in. This works much
better for that.